Rename unique index violation check to remove redundancy.
